Improve B -> C and C -> B transition glitch rejection
This sould also improve our handling of
IEC 61851-1 table A.6 sequence 11 "EV signal to the EV supply equipment".

+++ software/src/adc.h
@@ -68,7 +68,7 @@ typedef struct {
 } ADC;
 
 typedef struct {
-	uint8_t resistance_counter;
+	uint32_t resistance_counter;
 	uint32_t cp_pe_resistance;
 	uint32_t pp_pe_resistance;
 	int32_t cp_pe_pwm_low_mv[2];
